+++ src/extension/messageHandler.ts
@@ -352,7 +352,7 @@ export function registerMessageHandlers(
 
   registerAction("addRemote", (msg) => addRemote(gitClient.getInstance(), msg, recordGitCommand));
   registerAction("addTag", (msg) => addTag(gitClient.getInstance(), msg));
-  registerAction("deleteTag", (msg) => deleteTag(gitClient.getInstance(), msg));
+  registerAction("deleteTag", (msg) => deleteTag(gitClient.getInstance(), msg, recordGitCommand));
   registerAction("pushTag", (msg) => pushTag(gitClient.getInstance(), msg));
   registerAction("fetchRemotes", (msg) =>
     fetchRemotes(gitClient.getInstance(), msg, recordGitCommand)